[format] Apply yapf-0.40.0 globally

See merge request Toronto_DL_Lab/ncore!208

+++ ncore/impl/common/transformations.py
@@ -4,6 +4,7 @@ import numpy as np
 
 from scipy.spatial.transform import Rotation as R
 
+
 def so3_trans_2_se3(so3, trans):
     """Create a 4x4 rigid transformation matrix given so3 rotation and translation.
 
@@ -17,16 +18,20 @@ def so3_trans_2_se3(so3, trans):
 
     if so3.ndim > 2:
         T = np.eye(4)
-        T = np.tile(T,(so3.shape[0], 1, 1))
+        T = np.tile(T, (so3.shape[0], 1, 1))
         T[:, 0:3, 0:3] = so3
-        T[:, 0:3, 3] = trans.reshape(-1,3,)
+        T[:, 0:3, 3] = trans.reshape(
+            -1,
+            3,
+        )
 
     else:
         T = np.eye(4)
         T[0:3, 0:3] = so3
-        T[0:3, 3] = trans.reshape(3,)
+        T[0:3, 3] = trans.reshape(3, )
 
     return T
+
 
 def axis_angle_trans_2_se3(rot_axis, rot_angle, trans, degrees=True):
     ''' Converts the axis/angle rotation and translation to a se3 transformation matrix
@@ -41,8 +46,6 @@ def axis_angle_trans_2_se3(rot_axis, rot_angle, trans, degrees=True):
     '''
 
     return so3_trans_2_se3(axis_angle_2_so3(rot_axis, rot_angle, degrees), trans)
-
-
 
 
 def euler_trans_2_se3(euler_angles, trans, degrees=True, seq='xyz'):
@@ -60,7 +63,6 @@ def euler_trans_2_se3(euler_angles, trans, degrees=True, seq='xyz'):
     return so3_trans_2_se3(euler_2_so3(euler_angles, degrees), trans)
 
 
-
 def axis_angle_2_so3(axis, angle, degrees=True):
     ''' Converts the axis angle representation of the so3 rotation matrix
     Args:
@@ -76,7 +78,7 @@ def axis_angle_2_so3(axis, angle, degrees=True):
     angle[angle < cutoff_angle] = 0.0
 
     # Scale the axis to have the norm representing the angle
-    axis_angle = (angle/np.linalg.norm(axis, axis=1, keepdims=True)) * axis
+    axis_angle = (angle / np.linalg.norm(axis, axis=1, keepdims=True)) * axis
 
     return R.from_rotvec(axis_angle, degrees=degrees).as_matrix()
 
@@ -129,18 +131,20 @@ def so3_2_axis_angle(so3, degrees=True):
 def euclidean_2_spherical_coords(coords):
 
     r = np.linalg.norm(coords, axis=-1, keepdims=True)
-    el = np.arctan(coords[:,2]/np.linalg.norm(coords[:,:2], axis=-1)).reshape(-1,1)
-    az = np.arctan2(coords[:,1],coords[:,0]).reshape(-1,1)
+    el = np.arctan(coords[:, 2] / np.linalg.norm(coords[:, :2], axis=-1)).reshape(-1, 1)
+    az = np.arctan2(coords[:, 1], coords[:, 0]).reshape(-1, 1)
 
-    return np.concatenate((r,az,el),axis=-1)
+    return np.concatenate((r, az, el), axis=-1)
+
 
 def spherical_2_direction(spherical_coords):
 
-    dx = np.cos(spherical_coords[:,2]) * np.cos(spherical_coords[:,1])
-    dy = np.cos(spherical_coords[:,2]) * np.sin(spherical_coords[:,1])
-    dz = np.sin(spherical_coords[:,2])
+    dx = np.cos(spherical_coords[:, 2]) * np.cos(spherical_coords[:, 1])
+    dy = np.cos(spherical_coords[:, 2]) * np.sin(spherical_coords[:, 1])
+    dz = np.sin(spherical_coords[:, 2])
 
-    return np.concatenate((dx[:,None],dy[:,None],dz[:,None]),axis=-1)
+    return np.concatenate((dx[:, None], dy[:, None], dz[:, None]), axis=-1)
+
 
 def transform_point_cloud(pc, T):
     ''' Transform the point cloud with the provided transformation matrix, 
@@ -168,15 +172,14 @@ def transform_point_cloud(pc, T):
 def bbox_pose(bbox: np.ndarray) -> np.ndarray:
     ''' Converts an array-encoded bounding-box into a corresponding pose '''
 
-    return np.block(
-        [[R.from_euler('xyz', bbox[6:9], degrees=False).as_matrix(),
-          np.array(bbox[:3]).reshape((3, 1))], [np.array([0, 0, 0, 1])]])
+    return np.block([[R.from_euler('xyz', bbox[6:9], degrees=False).as_matrix(),
+                      np.array(bbox[:3]).reshape((3, 1))], [np.array([0, 0, 0, 1])]])
 
 
 def pose_bbox(pose: np.ndarray, dimensions: np.ndarray) -> np.ndarray:
     ''' Converts a pose with extents to an array-encoded bounding-box '''
 
-    bbox = np.empty(9, dtype = dimensions.dtype)
+    bbox = np.empty(9, dtype=dimensions.dtype)
     bbox[:3] = pose[:3, 3]  # centroid
     bbox[3:6] = dimensions  # dimensions from input
     bbox[6:9] = R.from_matrix(pose[:3, :3]).as_euler('xyz', degrees=False)  # orientation
@@ -212,7 +215,7 @@ def se3_inverse(T: np.ndarray) -> np.ndarray:
         (np array): inverse transformation [4,4]
     '''
     Rt = T[:3, :3].transpose()
-    return np.block([[Rt, - Rt @ T[:3,3:]], [np.zeros((1, 3)), 1]])
+    return np.block([[Rt, -Rt @ T[:3, 3:]], [np.zeros((1, 3)), 1]])
 
 
 def local_ENU_2_ECEF_orientation(theta, phi):
@@ -223,19 +226,19 @@ def local_ENU_2_ECEF_orientation(theta, phi):
     Out:
         (np.array): rotation from world pose to ECEF in so3 representation [n,3,3]
     '''
-    z_dir = np.concatenate([(np.sin(theta)*np.cos(phi))[:,None],
-                            (np.sin(theta)*np.sin(phi))[:,None],
-                            (np.cos(theta))[:,None] ],axis=1)
-    z_dir = z_dir/np.linalg.norm(z_dir, axis=-1, keepdims=True)
+    z_dir = np.concatenate([(np.sin(theta) * np.cos(phi))[:, None], (np.sin(theta) * np.sin(phi))[:, None],
+                            (np.cos(theta))[:, None]],
+                           axis=1)
+    z_dir = z_dir / np.linalg.norm(z_dir, axis=-1, keepdims=True)
 
-    y_dir = np.concatenate([-(np.cos(theta)*np.cos(phi))[:,None],
-                            -(np.cos(theta)*np.sin(phi))[:,None],
-                            (np.sin(theta))[:,None] ],axis=1)
-    y_dir = y_dir/np.linalg.norm(y_dir, axis=-1, keepdims=True)
+    y_dir = np.concatenate(
+        [-(np.cos(theta) * np.cos(phi))[:, None], -(np.cos(theta) * np.sin(phi))[:, None], (np.sin(theta))[:, None]],
+        axis=1)
+    y_dir = y_dir / np.linalg.norm(y_dir, axis=-1, keepdims=True)
 
     x_dir = np.cross(y_dir, z_dir)
 
-    return np.concatenate([x_dir[:,:,None], y_dir[:,:,None], z_dir[:,:,None]], axis = -1)
+    return np.concatenate([x_dir[:, :, None], y_dir[:, :, None], z_dir[:, :, None]], axis=-1)
 
 
 def lat_lng_alt_2_ECEF_elipsoidal(lat_lng_alt, a, b):
@@ -248,8 +251,8 @@ def lat_lng_alt_2_ECEF_elipsoidal(lat_lng_alt, a, b):
         (np.array): ECEF coordinates[n,3]
     '''
 
-    phi =  np.deg2rad(lat_lng_alt[:, 0])
-    gamma =  np.deg2rad(lat_lng_alt[:, 1])
+    phi = np.deg2rad(lat_lng_alt[:, 0])
+    gamma = np.deg2rad(lat_lng_alt[:, 1])
 
     cos_phi = np.cos(phi)
     sin_phi = np.sin(phi)
@@ -259,12 +262,12 @@ def lat_lng_alt_2_ECEF_elipsoidal(lat_lng_alt, a, b):
 
     N = a / np.sqrt(1 - e_square * sin_phi * sin_phi)
 
-
     x = (N + lat_lng_alt[:, 2]) * cos_phi * cos_gamma
     y = (N + lat_lng_alt[:, 2]) * cos_phi * sin_gamma
-    z = (N *  (b*b)/(a*a) + lat_lng_alt[:, 2]) * sin_phi
+    z = (N * (b * b) / (a * a) + lat_lng_alt[:, 2]) * sin_phi
 
-    return np.concatenate([x[:,None] ,y[:,None], z[:,None]], axis=1 )
+    return np.concatenate([x[:, None], y[:, None], z[:, None]], axis=1)
+
 
 def translation_2_lat_lng_alt_spherical(translation, earth_radius):
     ''' Computes the translation in the ECEF to latitude, longitude, altitude based on the spherical earth model
@@ -275,10 +278,11 @@ def translation_2_lat_lng_alt_spherical(translation, earth_radius):
         (np.array): latitude, longitude and altitude [n,3]
     '''
     altitude = np.linalg.norm(translation, axis=-1) - earth_radius
-    latitude = np.rad2deg(90 - np.arccos(translation[:,2] / np.linalg.norm(translation, axis=-1, keepdims=True)))
-    longitude =  np.rad2deg(np.arctan2(translation[:,1],translation[:,0]))
+    latitude = np.rad2deg(90 - np.arccos(translation[:, 2] / np.linalg.norm(translation, axis=-1, keepdims=True)))
+    longitude = np.rad2deg(np.arctan2(translation[:, 1], translation[:, 0]))
 
-    return np.concatenate([latitude[:,None], longitude[:,None], altitude[:,None]], axis=1)
+    return np.concatenate([latitude[:, None], longitude[:, None], altitude[:, None]], axis=1)
+
 
 def translation_2_lat_lng_alt_ellipsoidal(translation, a, f):
     ''' Computes the translation in the ECEF to latitude, longitude, altitude based on the ellipsoidal earth model
@@ -296,19 +300,21 @@ def translation_2_lat_lng_alt_ellipsoidal(translation, a, f):
     f1 = 1 - f0
     f2 = 1 / f0 - 1
 
-    z_div_1_f =  translation[:,2] / (1 - f)
-    x2y2 = np.square(translation[:,0]) + np.square(translation[:,1])
+    z_div_1_f = translation[:, 2] / (1 - f)
+    x2y2 = np.square(translation[:, 0]) + np.square(translation[:, 1])
 
-    x2y2z2 = x2y2 + z_div_1_f*z_div_1_f
+    x2y2z2 = x2y2 + z_div_1_f * z_div_1_f
     x2y2z2_pow_3_2 = x2y2z2 * np.sqrt(x2y2z2)
 
-    gamma = (x2y2z2_pow_3_2 + a * f2 * z_div_1_f * z_div_1_f) / (x2y2z2_pow_3_2 - a * f1 * x2y2) *  translation[:,2] / np.sqrt(x2y2)
+    gamma = (x2y2z2_pow_3_2 + a * f2 * z_div_1_f * z_div_1_f) / (x2y2z2_pow_3_2 -
+                                                                 a * f1 * x2y2) * translation[:, 2] / np.sqrt(x2y2)
 
-    longitude = np.rad2deg(np.arctan2(translation[:,1], translation[:,0]))
+    longitude = np.rad2deg(np.arctan2(translation[:, 1], translation[:, 0]))
     latitude = np.rad2deg(np.arctan(gamma))
     altitude = np.sqrt(1 + np.square(gamma)) * (np.sqrt(x2y2) - a / np.sqrt(1 + f0 * np.square(gamma)))
 
-    return np.concatenate([latitude[:,None], longitude[:,None], altitude[:,None]], axis=1)
+    return np.concatenate([latitude[:, None], longitude[:, None], altitude[:, None]], axis=1)
+
 
 def lat_lng_alt_2_ecef(lat_lng_alt, orientation_axis, orientation_angle, earth_model='WGS84'):
     ''' Computes the transformation from the world pose coordiante system to the earth centered earth fixed (ECEF) one
@@ -321,13 +327,12 @@ def lat_lng_alt_2_ecef(lat_lng_alt, orientation_axis, orientation_angle, earth_m
         trans (np.array): transformation parameters from world pose to ECEF coordinate system in se3 form (n, 4, 4)
     '''
     n = lat_lng_alt.shape[0]
-    trans = np.tile(np.eye(4).reshape(1,4,4),[n,1,1])
+    trans = np.tile(np.eye(4).reshape(1, 4, 4), [n, 1, 1])
 
     theta = np.deg2rad(90. - lat_lng_alt[:, 0])
     phi = np.deg2rad(lat_lng_alt[:, 1])
 
     R_enu_ecef = local_ENU_2_ECEF_orientation(theta, phi)
-
 
     if earth_model == 'WGS84':
         a = 6378137.0
@@ -336,20 +341,20 @@ def lat_lng_alt_2_ecef(lat_lng_alt, orientation_axis, orientation_angle, earth_m
         translation = lat_lng_alt_2_ECEF_elipsoidal(lat_lng_alt, a, b)
 
     elif earth_model == 'sphere':
-        earth_radius = 6378137.0 # Earth radius in meters
-        z_dir =  np.concatenate([(np.sin(theta)*np.cos(phi))[:,None],
-                            (np.sin(theta)*np.sin(phi))[:,None],
-                            (np.cos(theta))[:,None] ],axis=1)
+        earth_radius = 6378137.0  # Earth radius in meters
+        z_dir = np.concatenate([(np.sin(theta) * np.cos(phi))[:, None], (np.sin(theta) * np.sin(phi))[:, None],
+                                (np.cos(theta))[:, None]],
+                               axis=1)
 
-        translation = (earth_radius + lat_lng_alt[:, -1])[:,None] * z_dir
+        translation = (earth_radius + lat_lng_alt[:, -1])[:, None] * z_dir
 
     else:
-        raise ValueError ("Selected ellipsoid not implemented!")
+        raise ValueError("Selected ellipsoid not implemented!")
 
     world_pose_orientation = axis_angle_2_so3(orientation_axis, orientation_angle)
 
-    trans[:,:3,:3] =  R_enu_ecef @ world_pose_orientation
-    trans[:,:3,3] =  translation
+    trans[:, :3, :3] = R_enu_ecef @ world_pose_orientation
+    trans[:, :3, 3] = translation
 
     return trans
 
@@ -365,8 +370,8 @@ def ecef_2_lat_lng_alt(trans, earth_model='WGS84'):
         orientation_angle (np.array): orientation angle of the local ENU coordinate system in degrees [n,1]
     '''
 
-    translation = trans[:,:3,3]
-    rotation = trans[:,:3,:3]
+    translation = trans[:, :3, 3]
+    rotation = trans[:, :3, :3]
 
     if earth_model == 'WGS84':
         a = 6378137.0
@@ -374,26 +379,25 @@ def ecef_2_lat_lng_alt(trans, earth_model='WGS84'):
         lat_lng_alt = translation_2_lat_lng_alt_ellipsoidal(translation, a, flattening)
 
     elif earth_model == 'sphere':
-        earth_radius = 6378137.0 # Earth radius in meters
+        earth_radius = 6378137.0  # Earth radius in meters
         lat_lng_alt = translation_2_lat_lng_alt_spherical(translation, earth_radius)
 
     else:
-        raise ValueError ("Selected ellipsoid not implemented!")
-
+        raise ValueError("Selected ellipsoid not implemented!")
 
     # Compute the orientation axis and angle
     theta = np.deg2rad((90. - lat_lng_alt[:, 0]))
     phi = np.deg2rad(lat_lng_alt[:, 1])
 
-    R_ecef_enu = local_ENU_2_ECEF_orientation(theta, phi).transpose(0,2,1)
+    R_ecef_enu = local_ENU_2_ECEF_orientation(theta, phi).transpose(0, 2, 1)
 
     orientation = R_ecef_enu @ rotation
     orientation_axis, orientation_angle = so3_2_axis_angle(orientation)
 
-
     return lat_lng_alt, orientation_axis, orientation_angle
 
-def ecef_2_ENU(loc_ref_point: np.ndarray, earth_model: str ='WGS84'):
+
+def ecef_2_ENU(loc_ref_point: np.ndarray, earth_model: str = 'WGS84'):
     ''' 
     Compute the transformation matrix that transforms points from the ECEF to a local ENU coordinate frame
     Args:
@@ -410,25 +414,29 @@ def ecef_2_ENU(loc_ref_point: np.ndarray, earth_model: str ='WGS84'):
         a = 6378137.0
         flattening = 1.0 / 298.257223563
         b = a * (1.0 - flattening)
-        translation = lat_lng_alt_2_ECEF_elipsoidal(loc_ref_point, a, b).reshape(3,1)
+        translation = lat_lng_alt_2_ECEF_elipsoidal(loc_ref_point, a, b).reshape(3, 1)
 
     elif earth_model == 'sphere':
-        earth_radius = 6378137.0 # Earth radius in meters
-        z_dir =  np.concatenate([(np.sin(loc_ref_point[1])*np.cos(loc_ref_point[0]))[:,None],
-                            (np.sin(loc_ref_point[1])*np.sin(loc_ref_point[0]))[:,None],
-                            (np.cos(loc_ref_point[0]))[:,None] ],axis=1)
+        earth_radius = 6378137.0  # Earth radius in meters
+        z_dir = np.concatenate([(np.sin(loc_ref_point[1]) * np.cos(loc_ref_point[0]))[:, None],
+                                (np.sin(loc_ref_point[1]) * np.sin(loc_ref_point[0]))[:, None],
+                                (np.cos(loc_ref_point[0]))[:, None]],
+                               axis=1)
 
-        translation = ((earth_radius + loc_ref_point[:, -1])[:,None] * z_dir).reshape(3,1)
+        translation = ((earth_radius + loc_ref_point[:, -1])[:, None] * z_dir).reshape(3, 1)
 
     else:
-        raise ValueError ("Selected ellipsoid not implemented!")
+        raise ValueError("Selected ellipsoid not implemented!")
 
     rad_lat = np.deg2rad(loc_ref_point[:, 0])
     rad_lon = np.deg2rad(loc_ref_point[:, 1])
-    T_ecef_enu[:3,:3] = np.array([[-np.sin(rad_lon), np.cos(rad_lon), 0],
-                  [ -np.sin(rad_lat)*np.cos(rad_lon), -np.sin(rad_lat)*np.sin(rad_lon),  np.cos(rad_lat)],
-                  [  np.cos(rad_lat)*np.cos(rad_lon), np.cos(rad_lat)*np.sin(rad_lon),  np.sin(rad_lat)]])
+    T_ecef_enu[:3, :3] = np.array(
+        [[-np.sin(rad_lon), np.cos(rad_lon), 0],
+         [-np.sin(rad_lat) * np.cos(rad_lon), -np.sin(rad_lat) * np.sin(rad_lon),
+          np.cos(rad_lat)], [np.cos(rad_lat) * np.cos(rad_lon),
+                             np.cos(rad_lat) * np.sin(rad_lon),
+                             np.sin(rad_lat)]])
 
-    T_ecef_enu[:3,3:4] = -T_ecef_enu[:3,:3] @ translation
+    T_ecef_enu[:3, 3:4] = -T_ecef_enu[:3, :3] @ translation
 
     return T_ecef_enu
